refactor(edge): isolate watchdog coordination service boundaries
