+++ crates/edge/src/quic_listener/control_api/mod.rs
@@ -1,5 +1,3 @@
-use std::ffi::OsString;
-
 use hyper::{server::conn::http1, service::service_fn};
 use hyper_util::rt::TokioIo;
 use spooky_config::{
