+++ crates/edge/src/quic_listener/control_api/tests.rs
@@ -1,4 +1,4 @@
-use std::{collections::HashMap, path::Path, sync::Arc};
+use std::{collections::HashMap, ffi::OsString, path::Path, sync::Arc};
 
 use http_body_util::BodyExt;
 use log::LevelFilter;
@@ -122,8 +122,10 @@ fn runtime_bundle_control_api_state(
 
 #[test]
 fn watchdog_restart_env_keeps_path_when_present() {
-    let env =
-        QUICListener::watchdog_restart_env(Some(OsString::from("/usr/bin:/bin")), "timeout_spike");
+    let env = crate::watchdog::service::watchdog_restart_env(
+        Some(OsString::from("/usr/bin:/bin")),
+        "timeout_spike",
+    );
     let map: HashMap<OsString, OsString> = env.into_iter().collect();
 
     assert_eq!(
@@ -138,7 +140,7 @@ fn watchdog_restart_env_keeps_path_when_present() {
 
 #[test]
 fn watchdog_restart_env_omits_path_when_missing() {
-    let env = QUICListener::watchdog_restart_env(None, "poll_stall");
+    let env = crate::watchdog::service::watchdog_restart_env(None, "poll_stall");
     let map: HashMap<OsString, OsString> = env.into_iter().collect();
 
     assert!(!map.contains_key(&OsString::from("PATH")));
