+++ crates/edge/src/watchdog/mod.rs
@@ -1,3 +1,5 @@
 pub mod config;
 pub mod coordinator;
+pub mod service;
+pub mod state;
 pub mod time;
